Emacs: Switch from overlays to text properties
Overlays are generally easier and more proper to use for the purposes of log
analyzer UI functionality.  One of big advantages of overlays is that they,
unlike text properties, don't modify the text.  However overlays suffer from
poor performance and log analyzer UI uses them a lot, resulting in slow
movements over result file in ovirt-log-analyzer mode.

So let's replace most of the overlays with text properties, which are much
faster, and try to handle buffer modifications in a reasonable way.